+++ lib/admin.js
@@ -179,6 +179,12 @@ Admin.prototype.addUser = function(username, password, options, callback) {
   const args = Array.prototype.slice.call(arguments, 2);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
 
+  // Special case where there is no password ($external users)
+  if (typeof username === 'string' && password != null && typeof password === 'object') {
+    options = password;
+    password = null;
+  }
+
   options = args.length ? args.shift() : {};
   options = Object.assign({}, options);
   // Get the options
