+++ lib/db.js
@@ -14,7 +14,6 @@ const Logger = require('./core').Logger;
 const Collection = require('./collection');
 const mergeOptionsAndWriteConcern = require('./utils').mergeOptionsAndWriteConcern;
 const executeOperation = require('./utils').executeOperation;
-const applyWriteConcern = require('./utils').applyWriteConcern;
 const resolveReadPreference = require('./utils').resolveReadPreference;
 const ChangeStream = require('./change_stream');
 const deprecate = require('util').deprecate;
@@ -690,17 +689,7 @@ Db.prototype.dropCollection = function(name, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
   options = options || {};
 
-  // Command to execute
-  const cmd = { drop: name };
-
-  // Decorate with write concern
-  applyWriteConcern(cmd, { db: this }, options);
-
-  // options
-  const opts = Object.assign({}, this.s.options, { readPreference: ReadPreference.PRIMARY });
-  if (options.session) opts.session = options.session;
-
-  const dropCollectionOperation = new DropCollectionOperation(this, cmd, opts);
+  const dropCollectionOperation = new DropCollectionOperation(this, name, options);
 
   return executeOperationV2(this.s.topology, dropCollectionOperation, callback);
 };
@@ -717,22 +706,8 @@ Db.prototype.dropCollection = function(name, options, callback) {
 Db.prototype.dropDatabase = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
   options = options || {};
-  // Drop database command
-  const cmd = { dropDatabase: 1 };
 
-  // Decorate with write concern
-  applyWriteConcern(cmd, { db: this }, options);
-
-  // Ensure primary only
-  const finalOptions = Object.assign({}, this.s.options, {
-    readPreference: ReadPreference.PRIMARY
-  });
-
-  if (options.session) {
-    finalOptions.session = options.session;
-  }
-
-  const dropDatabaseOperation = new DropDatabaseOperation(this, cmd, finalOptions);
+  const dropDatabaseOperation = new DropDatabaseOperation(this, options);
 
   return executeOperationV2(this.s.topology, dropDatabaseOperation, callback);
 };
@@ -870,6 +845,12 @@ Db.prototype.addChild = function(db) {
 Db.prototype.addUser = function(username, password, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
   options = options || {};
+
+  // Special case where there is no password ($external users)
+  if (typeof username === 'string' && password != null && typeof password === 'object') {
+    options = password;
+    password = null;
+  }
 
   const addUserOperation = new AddUserOperation(this, username, password, options);
