+++ lib/operations/add_user.js
@@ -15,14 +15,6 @@ const WriteConcern = require('../write_concern');
 
 class AddUserOperation extends CommandOperation {
   constructor(db, username, password, options) {
-    options = options || {};
-
-    // Special case where there is no password ($external users)
-    if (typeof username === 'string' && password != null && typeof password === 'object') {
-      options = password;
-      password = null;
-    }
-
     super(db, {}, options);
 
     this.username = username;
